Harden deployment scripts and add EDRR workflow tests
