+++ scripts/deployment/stop_stack.sh
@@ -7,4 +7,23 @@ set -euo pipefail
 
 ENVIRONMENT=${1:-development}
 
-docker compose --profile "${ENVIRONMENT}" --profile monitoring down
+# Enforce non-root execution for least privilege
+if [[ "$EUID" -eq 0 ]]; then
+  echo "Please run this script as a non-root user." >&2
+  exit 1
+fi
+
+# Allow only known environments
+ALLOWED_ENVIRONMENTS=(development staging production testing)
+if [[ ! " ${ALLOWED_ENVIRONMENTS[*]} " =~ " ${ENVIRONMENT} " ]]; then
+  echo "Invalid environment: ${ENVIRONMENT}" >&2
+  exit 1
+fi
+
+ENV_FILE=".env.${ENVIRONMENT}"
+if [[ -f "$ENV_FILE" ]] && [[ $(stat -c %a "$ENV_FILE") != "600" ]]; then
+  echo "Environment file $ENV_FILE must have 600 permissions" >&2
+  exit 1
+fi
+
+docker compose --env-file "$ENV_FILE" --profile "${ENVIRONMENT}" --profile monitoring down
